requirements.txt
streamlit
langchain
langchain-community
langchain-groq
beautifulsoup4
python-dotenv